set up conditional redirect when creating a trail
When a user visits the new trail form directly and submits the form, the user is redirected to the show page for that trail. When a user visits the create event page and follows the link from there for creating a new trail (and subsequently submits the new trail form), the user is redirected to the new event page. The event being created on that page belongs to the trail that was just created.